moved EdifactValidationException into exceptions package
