+++ src/main/java/uk/nhs/digital/nhsconnect/nhais/model/edifact/MessageHeader.java
@@ -3,6 +3,7 @@ package uk.nhs.digital.nhsconnect.nhais.model.edifact;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
+import uk.nhs.digital.nhsconnect.nhais.exceptions.EdifactValidationException;
 
 /**
  * A specialisation of a segment for the specific use case of a message header
