+++ src/main/java/uk/nhs/digital/nhsconnect/nhais/model/edifact/NameAndAddress.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
+import uk.nhs.digital.nhsconnect.nhais.exceptions.EdifactValidationException;
 
 /**
  *class declaration:
