+++ src/main/java/uk/nhs/digital/nhsconnect/nhais/model/edifact/ReferenceTransactionNumber.java
@@ -3,6 +3,7 @@ package uk.nhs.digital.nhsconnect.nhais.model.edifact;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.Setter;
+import uk.nhs.digital.nhsconnect.nhais.exceptions.EdifactValidationException;
 
 /**
  *class declaration:
