+++ src/test/java/uk/nhs/digital/nhsconnect/nhais/model/edifact/NameAndAddressTest.java
@@ -1,6 +1,7 @@
 package uk.nhs.digital.nhsconnect.nhais.model.edifact;
 
 import org.junit.jupiter.api.Test;
+import uk.nhs.digital.nhsconnect.nhais.exceptions.EdifactValidationException;
 
 import static org.junit.jupiter.api.Assertions.*;
